Implementation der Model Klassen für das Passwortspiel

+++ app/src/main/java/com/securify/securify/gameModels/GameModel.java
@@ -19,7 +19,9 @@ public class GameModel {
     private int zeit;
     private String erklaerung;
 
-    public void setId(int id) {
+
+
+    public void setId(long id) {
         this.id = id;
     }
